lac and waybill use int64 for windows compatibility

+++ examples/lexical_analysis/predict.py
@@ -46,8 +46,8 @@ def infer(args):
     infer_dataset = LacDataset(args.data_dir, mode='infer')
 
     batchify_fn = lambda samples, fn=Tuple(
-        Pad(axis=0, pad_val=0),  # word_ids
-        Stack(),  # length
+        Pad(axis=0, pad_val=0, dtype='int64'),  # word_ids
+        Stack(dtype='int64'),  # length
     ): fn(samples)
 
     # Create sampler for dataloader
